# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beeba Boys Barbershop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: string;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  duration_minutes?: number | null;
  /** @nullable */
  price?: number | null;
  active: boolean;
}

export interface ServiceInput {
  name: string;
  category: string;
  description?: string;
  duration_minutes?: number;
  price?: number;
  active?: boolean;
}

export interface ServiceUpdate {
  name?: string;
  category?: string;
  description?: string;
  duration_minutes?: number;
  price?: number;
  active?: boolean;
}

export interface TeamMember {
  id: string;
  name: string;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  speciality?: string | null;
  /** @nullable */
  photo_url?: string | null;
  /** @nullable */
  years_experience?: number | null;
  active: boolean;
}

export interface TeamMemberInput {
  name: string;
  role: string;
  bio?: string;
  speciality?: string;
  photo_url?: string;
  years_experience?: number;
  active?: boolean;
}

export interface TeamMemberUpdate {
  name?: string;
  role?: string;
  bio?: string;
  speciality?: string;
  photo_url?: string;
  years_experience?: number;
  active?: boolean;
}

export interface GalleryImage {
  id: string;
  image_url: string;
  category: string;
  created_at?: string;
}

export interface GalleryImageInput {
  image_url: string;
  category: string;
}

export interface Booking {
  id: string;
  client_name: string;
  /** @nullable */
  client_email?: string | null;
  /** @nullable */
  client_phone?: string | null;
  /** @nullable */
  service_id?: string | null;
  /** @nullable */
  stylist_id?: string | null;
  /** @nullable */
  appointment_date?: string | null;
  /** @nullable */
  appointment_time?: string | null;
  status: string;
  created_at: string;
  /** @nullable */
  service_name?: string | null;
  /** @nullable */
  stylist_name?: string | null;
}

export interface BookingInput {
  client_name: string;
  client_email?: string;
  client_phone?: string;
  service_id?: string;
  stylist_id?: string;
  appointment_date?: string;
  appointment_time?: string;
}

export interface BookingUpdate {
  status?: string;
}

export interface Message {
  id: string;
  name: string;
  email: string;
  message: string;
  read: boolean;
  created_at: string;
}

export interface MessageInput {
  name: string;
  email: string;
  message: string;
}

export interface MessageUpdate {
  read?: boolean;
}

export interface Setting {
  key: string;
  value: string;
}

export interface SettingsInput {
  settings: Setting[];
}

export interface DashboardStats {
  total_bookings: number;
  pending_bookings: number;
  confirmed_bookings: number;
  unread_messages: number;
  total_services: number;
  total_team: number;
}