# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beeba Boys Barbershop API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: string;
  client_name: string;
  /** @nullable */
  client_email?: string | null;
  /** @nullable */
  client_phone?: string | null;
  /** @nullable */
  service_id?: string | null;
  /** @nullable */
  stylist_id?: string | null;
  /** @nullable */
  appointment_date?: string | null;
  /** @nullable */
  appointment_time?: string | null;
  status: string;
  created_at: string;
  /** @nullable */
  service_name?: string | null;
  /** @nullable */
  stylist_name?: string | null;
}