# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beeba Boys Barbershop API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  client_name: string;
  client_email?: string;
  client_phone?: string;
  service_id?: string;
  stylist_id?: string;
  appointment_date?: string;
  appointment_time?: string;
}