# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beeba Boys Barbershop API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  name: string;
  email: string;
  message: string;
}