# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beeba Boys Barbershop API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  duration_minutes?: number | null;
  /** @nullable */
  price?: number | null;
  active: boolean;
}