# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beeba Boys Barbershop API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  name?: string;
  category?: string;
  description?: string;
  duration_minutes?: number;
  price?: number;
  active?: boolean;
}