# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beeba Boys Barbershop API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMember {
  id: string;
  name: string;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  speciality?: string | null;
  /** @nullable */
  photo_url?: string | null;
  /** @nullable */
  years_experience?: number | null;
  active: boolean;
}