# lib/api-zod/src/generated/types/teamMemberUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beeba Boys Barbershop API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMemberUpdate {
  name?: string;
  role?: string;
  bio?: string;
  speciality?: string;
  photo_url?: string;
  years_experience?: number;
  active?: boolean;
}